weekend02 algorithm - kangaroo First Refactoring
