+++ algorithm_project/src/main/java/com/example/kangaroo/model/Kangaroo.java
@@ -1,0 +1,6 @@
+package com.example.kangaroo.model;
+
+public class Kangaroo {
+	private int position;
+	private int distance;
+}
