+++ algorithm_project/src/main/java/com/example/kangaroo/service/KangarooService.java
@@ -1,0 +1,7 @@
+package com.example.kangaroo.service;
+
+import com.example.kangaroo.model.Kangaroo;
+
+public interface KangarooService {
+	String jumpKangaroo(Kangaroo firstKangaroo, Kangaroo secondKangaroo);
+}
